update: 更新文件 @ohos.multimodalAwareness.deviceStatus.d.ts 修改DeviceRotationRadian里面的类型从float为double

+++ api/@ohos.multimodalAwareness.deviceStatus.d.ts
@@ -68,31 +68,31 @@ declare namespace deviceStatus {
   export interface DeviceRotationRadian {
     /**
      * indicates X-RotationRadian
-     * @type { float }
+     * @type { double }
      * @syscap SystemCapability.MultimodalAwareness.DeviceStatus
      * @systemapi
      * @since 20 dynamic
      * @since 23 static
      */
-    x: float;
+    x: double;
     /**
      * indicates Y-RotationRadian
-     * @type { float }
+     * @type { double }
      * @syscap SystemCapability.MultimodalAwareness.DeviceStatus
      * @systemapi
      * @since 20 dynamic
      * @since 23 static
      */
-    y: float;
+    y: double;
     /**
      * indicates Z-RotationRadian
-     * @type { float }
+     * @type { double }
      * @syscap SystemCapability.MultimodalAwareness.DeviceStatus
      * @systemapi
      * @since 20 dynamic
      * @since 23 static
      */
-    z: float;
+    z: double;
   }
 
   /**
